Add ListView to Demo and extract the functional part out to a separate DLL packaged ready for publish on NuGet.
