+++ SelectedItemsBindingDemo/App.xaml
