+++ SelectedItemsBindingDemo/App.xaml.cs
