+++ SelectedItemsBindingDemo/Properties/AssemblyInfo.cs
