+++ SelectedItemsBindingDemo/Properties/Resources.resx
