+++ SelectedItemsBindingDemo/Properties/Settings.settings
