+++ SelectedItemsBindingDemo/RelayCommand.cs
