+++ SelectedItemsBindingDemo/ViewModel.cs
