+++ SelectedItemsBindingDemo/Window1.xaml
@@ -1,0 +1,46 @@
+﻿<Window x:Class="SelectedItemsBindingDemo.Window1"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:d="clr-namespace:SelectedItemsBindingDemo"
+    xmlns:ffx="clr-namespace:PrimS.SelectedItemsSynchronizer;assembly=PrimS.SelectedItemsSynchronizer"
+    xmlns:i="http://schemas.microsoft.com/expression/2010/interactivity"
+    Title="SelectedItems Binding Demo" Height="300" Width="300">
+  <Window.DataContext>
+    <d:ViewModel/>
+  </Window.DataContext>
+  <Grid Margin="5">
+    <Grid.RowDefinitions>
+      <RowDefinition Height="*"/>
+      <RowDefinition Height="*"/>
+      <RowDefinition Height="Auto"/>
+    </Grid.RowDefinitions>
+    <TextBlock Grid.Row="2" HorizontalAlignment="Left" VerticalAlignment="Center" 
+                 Text="{Binding Summary}" FontWeight="Bold" Margin ="0,5,5,5"/>
+    <Button Content="Select All" Command="{Binding SelectAll}" Grid.Row="2" 
+              HorizontalAlignment="Right" VerticalAlignment="Center"/>
+    <ListBox x:Name="NamesListBox" 
+               Grid.Row="0" 
+               ItemsSource="{Binding AvailableNames}"
+               ffx:MultiSelectorBehaviours.SynchronizedSelectedItems="{Binding SelectedNames}"
+               SelectionMode="Extended">
+      <i:Interaction.Triggers>
+        <i:EventTrigger EventName="SelectionChanged">
+          <i:InvokeCommandAction Command="{Binding SelectionChangedCommand}" 
+                        CommandParameter="{Binding ElementName=NamesListBox, Path=SelectedItems}" />
+        </i:EventTrigger>
+      </i:Interaction.Triggers>
+    </ListBox>
+    <ListView x:Name="NamesListView"
+              Grid.Row="1"
+              ItemsSource="{Binding AvailableNames}"
+              ffx:MultiSelectorBehaviours.SynchronizedSelectedItems="{Binding SelectedNames}"
+              SelectionMode="Extended">
+      <i:Interaction.Triggers>
+        <i:EventTrigger EventName="SelectionChanged">
+          <i:InvokeCommandAction Command="{Binding SelectionChangedCommand}" 
+                        CommandParameter="{Binding ElementName=NamesListBox, Path=SelectedItems}" />
+        </i:EventTrigger>
+      </i:Interaction.Triggers>
+    </ListView>
+  </Grid>
+</Window>
