+++ SelectedItemsBindingDemo/Window1.xaml.cs
