+++ SelectedItemsBindingDemo/app.config
