+++ SelectedItemsSynchronizer.nuspec
@@ -1,0 +1,25 @@
+<?xml version="1.0"?>
+<package >
+  <metadata>
+    <id>SelectedItemsSynchronizer</id>
+    <version>0.5.0-beta</version>
+    <title>SelectedItemsSynchronizer</title>
+    <authors>SelectedItemsSynchronizer contributors</authors>
+    <owners>9swampy</owners>
+    <licenseUrl>https://github.com/9swampy/SelectedItemsSynchronizer/blob/SelectedItemsSynchronizer/License.txt</licenseUrl>
+    <projectUrl>https://github.com/9swampy/SelectedItemsSynchronizer</projectUrl>
+    <iconUrl>http://www.prims.com/P128.png</iconUrl>
+    <requireLicenseAcceptance>false</requireLicenseAcceptance>
+    <description>Facilitates binding WPF SelectedItems</description>
+    <releaseNotes>
+      Beta release.
+    </releaseNotes>
+    <copyright>Copyright (c) 9swampy, SelectedItemSynchronizer contributors</copyright>
+    <tags>SelectedItems WPF binding ViewModel MVVM</tags>
+  </metadata>
+  <files>
+    <file src="License.txt" target="content" />
+    <file src="SelectedItemsSynchronizer\Bin\Release\PrimS.SelectedItemsSynchronizer.dll" target="lib\net45" />
+    <file src="SelectedItemsSynchronizer\Bin\Release\PrimS.SelectedItemsSynchronizer.xml" target="lib\net45" />
+  </files>
+</package>
