+++ SelectedItemsSynchronizer.sln
@@ -1,9 +1,11 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 2013
-VisualStudioVersion = 12.0.30501.0
+VisualStudioVersion = 12.0.30723.0
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SelectedItemsBindingDemo", "SelectedItemsBindingDemo.csproj", "{7DA5A107-B474-4AC0-B861-63A489DB0C02}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SelectedItemsBindingDemo", "SelectedItemsBindingDemo\SelectedItemsBindingDemo.csproj", "{7DA5A107-B474-4AC0-B861-63A489DB0C02}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PrimS.SelectedItemsSynchronizer", "SelectedItemsSynchronizer\PrimS.SelectedItemsSynchronizer.csproj", "{6FD3314F-9A43-4FCF-A073-6C97CA2D1E81}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -15,6 +17,10 @@ Global
 		{7DA5A107-B474-4AC0-B861-63A489DB0C02}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{7DA5A107-B474-4AC0-B861-63A489DB0C02}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{7DA5A107-B474-4AC0-B861-63A489DB0C02}.Release|Any CPU.Build.0 = Release|Any CPU
+		{6FD3314F-9A43-4FCF-A073-6C97CA2D1E81}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{6FD3314F-9A43-4FCF-A073-6C97CA2D1E81}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{6FD3314F-9A43-4FCF-A073-6C97CA2D1E81}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{6FD3314F-9A43-4FCF-A073-6C97CA2D1E81}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
