+++ SelectedItemsSynchronizer/packages.config
@@ -1,0 +1,4 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="StyleCop.MSBuild" version="4.7.49.0" targetFramework="net45" />
+</packages>
